refactor(favorites): decompose page directory - move EmptyFavorites, FavoritesFilter into components directory - create common barrel file
