+++ src/pages/Favorites/Favorites.module.css
@@ -11,86 +11,9 @@
   color: var(--color-text-primary);
 }
 
-.filterWrapper {
-  width: 600px;
-  position: relative;
-}
-
-.filter {
-  width: 100%;
-  padding: 0.65rem 2.5rem;
-  border: 1px solid #bbbbbb;
-  border-radius: 8px;
-  font-size: 16px;
-  color: var(--color-text-secondary);
-  text-overflow: ellipsis;
-}
-
-.filter:focus {
-  outline: 1px solid var(--color-primary);
-  border: 1px solid var(--color-primary);
-}
-
-.filterIcon,
-.clearButton {
-  position: absolute;
-  top: 10px;
-}
-
-.filterIcon {
-  left: 10px;
-}
-
-.clearButton {
-  right: 10px;
-}
-
 .productsGrid {
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
   gap: 1rem;
   padding: 0.5rem 0;
-}
-
-.empty {
-  height: 60vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.emptyImage {
-  width: clamp(350px, 50vw, 500px);
-}
-
-.message {
-  color: #1d5798;
-  font-weight: 600;
-  font-size: clamp(1.25rem, 2vw + 0.5rem, 1.75rem);
-  text-align: center;
-}
-
-.submessage {
-  color: #3b6ba3;
-  font-size: 1rem;
-  width: 60%;
-  text-align: center;
-}
-
-@media (max-width: 630px) {
-  .filterWrapper {
-    width: 100%;
-  }
-}
-
-@media (max-width: 600px) {
-  .empty {
-    height: 50vh;
-  }
-
-  .submessage {
-    width: 100%;
-  }
 }
