+++ src/pages/Favorites/components/EmptyFavorites/EmptyFavorites.module.css
@@ -1,0 +1,36 @@
+.empty {
+  height: 60vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.emptyImage {
+  width: clamp(350px, 50vw, 500px);
+}
+
+.message {
+  color: #1d5798;
+  font-weight: 600;
+  font-size: clamp(1.25rem, 2vw + 0.5rem, 1.75rem);
+  text-align: center;
+}
+
+.submessage {
+  color: #3b6ba3;
+  font-size: 1rem;
+  width: 60%;
+  text-align: center;
+}
+
+@media (max-width: 600px) {
+  .empty {
+    height: 50vh;
+  }
+
+  .submessage {
+    width: 100%;
+  }
+}
