+++ src/pages/Favorites/components/index.ts
@@ -1,0 +1,3 @@
+export { default as EmptyFavorites } from './EmptyFavorites/EmptyFavorites';
+export { default as FavoritesFilter } from './FavoritesFilter/FavoritesFilter';
+export { default as CategoryFilter } from './CategoryFilter/CategoryFilter';
